power

+++ 68714/power/attack.c
@@ -288,9 +288,9 @@ void attack() {
         //Calculate Sample Standard Deviation
         sum_H=0; sum_T=0;
         for (int i = 0;i<M_SIZE;i++){
-          sum_H +=(h[i][ki] - mean_H);
-          sum_T +=(t[i][j] - mean_T);
-          printf("%f %f\n", sum_H, h[i][ki]);
+          sum_H +=((double)h[i][ki] - mean_H);
+          sum_T +=((double)t[i][j] - mean_T);
+          printf("%f %f\n", sum_H, (double)h[i][ki]);
         }
         double s_H = sqrt(sum_H/(double)(M_SIZE-1));
         double s_T = sqrt(sum_T/(double)(M_SIZE-1));
